Change the current directory when running terraform
Some tools (e.g. tfenv) wrap the terraform executable and rely on the
execution context to find the actual terraform binary. For tfenv to
pick up the right terraform version it needs to be run in the project
folder.

The current directory is set to the directory of the file that is being
formatted. This ensures that terraform is always run in the right
context.